requirements.txt
moviepy==1.0.3
pydub==0.25.1
python-dotenv==1.0.0
